[Refactor/#84] tutorial overlay 재사용 가능하게 분리
