+++ app/src/main/res/drawable/ic_schedule_tooltip.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="193dp"
+    android:height="52dp"
+    android:viewportWidth="193"
+    android:viewportHeight="52">
+  <path
+      android:pathData="M0,8C0,3.58 3.58,0 8,0H185C189.42,0 193,3.58 193,8V32C193,36.42 189.42,40 185,40H8C3.58,40 0,36.42 0,32V8Z"
+      android:fillColor="#FF6A3A"/>
+  <path
+      android:pathData="M19.13,25.73V13.65H20.67V25.73H19.13ZM18.27,22.17V23.4H10.59V22.17H13.7V20.37H15.24V22.17H18.27ZM11.72,20.76L10.74,19.72L13.11,17.47C13.54,17.08 13.68,16.77 13.71,16.45H10.95V15.19H13.74V13.83H15.24V15.19H18.04V16.45H15.26C15.31,16.77 15.45,17.1 15.85,17.47L18.23,19.72L17.25,20.76L14.9,18.5C14.75,18.36 14.61,18.21 14.48,18.07C14.37,18.21 14.23,18.36 14.06,18.5L11.72,20.76ZM31.87,13.66H33.34V25.7H31.87V19.79H30.51V25.08H29.09V13.91H30.51V18.53H31.87V13.66ZM28.2,21.23V22.51H23.22V14.74H27.92V16H24.76V21.23H28.2ZM42.29,25.13C40.31,25.13 38.84,23.82 38.84,22.07H40.33C40.33,23.05 41.17,23.81 42.29,23.81C43.52,23.81 44.44,22.96 44.44,21.84C44.44,20.73 43.53,19.89 42.31,19.89H41.87L40.84,18.91L43.45,16.12H39.33V14.81H45.21L45.35,16.12L43.01,18.64C44.69,18.92 45.88,20.2 45.88,21.84C45.88,23.71 44.34,25.13 42.29,25.13ZM51.91,23.82C53.69,23.82 54.81,22.33 54.81,19.95C54.81,17.55 53.69,16.05 51.91,16.05C50.12,16.05 49,17.55 49,19.95C49,22.33 50.12,23.82 51.91,23.82ZM51.91,25.14C49.21,25.14 47.52,23.12 47.52,19.95C47.52,16.75 49.21,14.75 51.91,14.75C54.6,14.75 56.29,16.75 56.29,19.95C56.29,23.12 54.6,25.14 51.91,25.14ZM66.37,13.67H67.91V19.96H66.37V13.67ZM60.04,20.51H67.91V23.77H61.57V24.78H68.19V25.98H60.07V22.65H66.38V21.7H60.04V20.51ZM61.16,19.92C59.26,19.92 57.93,18.67 57.93,16.92C57.93,15.16 59.26,13.93 61.16,13.93C63.05,13.93 64.38,15.16 64.38,16.92C64.38,18.67 63.05,19.92 61.16,19.92ZM61.16,18.7C62.18,18.7 62.87,17.99 62.87,16.92C62.87,15.86 62.18,15.13 61.16,15.13C60.13,15.13 59.44,15.86 59.44,16.92C59.44,17.99 60.13,18.7 61.16,18.7ZM81.67,13.66H83.21V16.38H84.95V17.64H83.21V20.7H81.67V13.66ZM75.47,21.3H83.21V25.97H75.47V21.3ZM81.71,22.51H76.98V24.75H81.71V22.51ZM80.4,18.77V20.03H74.03V14.32H75.57V18.77H80.4ZM85.78,20.02H97.03V21.25H85.78V20.02ZM87.51,22.19H89.05V24.65H95.49V25.9H87.51V22.19ZM88,16.36C88,14.7 89.39,13.6 91.38,13.6C93.39,13.6 94.81,14.7 94.81,16.36C94.81,17.93 93.39,19.09 91.38,19.09C89.39,19.09 88,17.93 88,16.36ZM89.47,16.36C89.47,17.2 90.24,17.83 91.39,17.83C92.55,17.83 93.34,17.2 93.34,16.36C93.34,15.41 92.55,14.85 91.39,14.85C90.24,14.85 89.47,15.41 89.47,16.36ZM110.36,13.67H111.9V19.96H110.36V13.67ZM104.03,20.51H111.9V23.77H105.56V24.78H112.18V25.98H104.06V22.65H110.37V21.7H104.03V20.51ZM105.15,19.92C103.25,19.92 101.92,18.67 101.92,16.92C101.92,15.16 103.25,13.93 105.15,13.93C107.04,13.93 108.37,15.16 108.37,16.92C108.37,18.67 107.04,19.92 105.15,19.92ZM105.15,18.7C106.17,18.7 106.86,17.99 106.86,16.92C106.86,15.86 106.17,15.13 105.15,15.13C104.11,15.13 103.43,15.86 103.43,16.92C103.43,17.99 104.11,18.7 105.15,18.7ZM120.71,16.54H122.74V13.67H124.28V21.01H122.74V17.8H120.71V16.54ZM115.31,20.25L114.27,19.27L116.43,16.96C116.86,16.5 116.96,16.14 116.96,15.77V15.49H114.55V14.23H120.71V15.49H118.43V15.77C118.43,16.17 118.5,16.5 118.94,16.96L121.11,19.27L120.08,20.25L117.91,17.94L117.68,17.69L117.46,17.94L115.31,20.25ZM120.56,26.39C118.69,26.39 117.39,25.27 117.39,23.7C117.39,22.12 118.69,21 120.56,21C122.43,21 123.75,22.12 123.75,23.7C123.75,25.27 122.43,26.39 120.56,26.39ZM120.56,25.14C121.56,25.14 122.25,24.55 122.25,23.7C122.25,22.83 121.56,22.26 120.56,22.26C119.56,22.26 118.89,22.83 118.89,23.7C118.89,24.55 119.56,25.14 120.56,25.14ZM126.69,20.34V14.5H132.54V20.34H126.69ZM128.2,21.68H129.77V24.69H136.32V25.94H128.2V21.68ZM128.2,19.11H131.03V15.73H128.2V19.11ZM134.48,22.24V13.66H136.02V17.33H137.74V18.6H136.02V22.24H134.48ZM142.09,22.94H146.95V20.76H143.7V14.4H145.23V16.36H150.22V14.4H151.74V20.76H148.49V22.94H153.37V24.2H142.09V22.94ZM145.23,17.58V19.54H150.22V17.58H145.23ZM163.25,13.65H164.77V25.73H163.25V20.74H160.41C159.8,21.44 158.91,21.85 157.86,21.85C155.87,21.85 154.46,20.39 154.46,18.32C154.46,16.24 155.87,14.77 157.86,14.77C158.96,14.77 159.89,15.23 160.49,15.98H163.25V13.65ZM157.86,20.53C158.99,20.53 159.75,19.64 159.75,18.32C159.75,16.99 158.99,16.08 157.86,16.08C156.72,16.08 155.97,16.99 155.97,18.32C155.97,19.64 156.72,20.53 157.86,20.53ZM161.08,19.5H163.25V17.23H161.11C161.2,17.57 161.25,17.93 161.25,18.32C161.25,18.74 161.19,19.13 161.08,19.5ZM166.74,23.08H169.69V20.27C168.99,19.64 168.57,18.71 168.57,17.64C168.57,15.54 170.14,14.05 172.35,14.05C174.61,14.05 176.18,15.54 176.18,17.64C176.18,18.7 175.77,19.62 175.07,20.25V23.08H178.02V24.33H166.74V23.08ZM170.09,17.64C170.09,18.98 170.99,19.95 172.35,19.95C173.74,19.95 174.66,18.98 174.66,17.64C174.66,16.29 173.74,15.33 172.35,15.33C170.99,15.33 170.09,16.29 170.09,17.64ZM171.22,23.08H173.53V21.07C173.16,21.16 172.77,21.22 172.35,21.22C171.95,21.22 171.57,21.18 171.22,21.08V23.08ZM180.31,23.59C180.81,23.59 181.16,23.95 181.16,24.45C181.16,24.97 180.81,25.34 180.31,25.34C179.79,25.34 179.43,24.97 179.43,24.45C179.43,23.95 179.79,23.59 180.31,23.59Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M153,52L146.07,40L159.93,40L153,52Z"
+      android:fillColor="#FF6A3A"/>
+</vector>
